requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
kubernetes>=27.2.0
PyYAML>=6.0
requests>=2.31.0
